// tc-be/src/common/interfaces/multer-file.interface.ts
// Fix Express.Multer type issue by using correct type import
export interface IMulterFileInterface {
    fieldname: string;
    originalname: string;
    encoding: string;
    mimetype: string;
    size: number;
    destination: string;
    filename: string;
    path: string;
    buffer: Buffer;
}